requirements.txt
numpy==1.26.3
scipy==1.11.3
scikit-learn==1.3.2
pydub==0.25.1
Pillow==10.0.1
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.2
librosa==0.10.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
urllib3==1.26.18
Werkzeug==3.0.0
Flask==3.0.3
zipp==3.17.0